requirements.txt
pandas
numpy
matplotlib
Pillow
torch
torch-vision